Expose couple helper functions
